refactor(desktop): split agent chat message card model utilities
Extract formatting, lifecycle, question parsing, and summary helpers into focused modules while keeping agent-chat-message-card-model.ts as a compatibility barrel export.
